requirements.txt
elasticsearch==8.13.1
Flask==3.0.3
python-dotenv==1.0.1
sentence_transformers==3.0.0
